+++ src/components/pbmg.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
 
 export const PoweredByMapGuide = (props: any) => {
-    return <img {...props} src="stdicons/PoweredBy_en.gif" />;
+    return <img className="component-pbmg" {...props} src="stdicons/PoweredBy_en.gif" />;
 };