+++ src/layouts/ajax-viewer.tsx
@@ -12,22 +12,6 @@ import {
 const SIDEBAR_WIDTH = 250;
 const LEGEND_HEIGHT = 350;
 
-const NAVIGATOR_PROPS = {
-    style: { position: "absolute", zIndex: 1000, width: 51, height: 204, cursor: "pointer", right: 10, top: 10 }
-};
-const MOUSE_COORDINATE_PROPS = {
-    style: { position: "absolute", bottom: 0, left: 0, zIndex: 100, backgroundColor: TOOLBAR_BACKGROUND_COLOR }
-};
-const SCALE_DISPLAY_PROPS = {
-    style: { position: "absolute", bottom: 0, right: 340, zIndex: 100, backgroundColor: TOOLBAR_BACKGROUND_COLOR }
-};
-const SELECTED_FEATURE_COUNT_PROPS = {
-    style: { position: "absolute", bottom: 0, right: 140, zIndex: 100, backgroundColor: TOOLBAR_BACKGROUND_COLOR }
-};
-const PBMG_PROPS = {
-    style: { position: "absolute", bottom: 0, right: 0, zIndex: 100 }
-};
-
 export interface IAjaxViewerLayoutState {
     map?: any;
     config?: any;
@@ -127,25 +111,25 @@ export class AjaxViewerLayout extends React.Component<AjaxViewerLayoutProps, any
                     })()}
                     {(() => {
                         if (hasNavigator) {
-                            return <PlaceholderComponent id={DefaultComponentNames.Navigator} locale={this.props.config.locale} componentProps={NAVIGATOR_PROPS} />;
+                            return <PlaceholderComponent id={DefaultComponentNames.Navigator} locale={this.props.config.locale} />;
                         }
                     })()}
                     {(() => {
                         if (hasStatusBar) {
-                            return <PlaceholderComponent id={DefaultComponentNames.MouseCoordinates} locale={this.props.config.locale} componentProps={MOUSE_COORDINATE_PROPS} />;
+                            return <PlaceholderComponent id={DefaultComponentNames.MouseCoordinates} locale={this.props.config.locale} />;
                         }
                     })()}
                     {(() => {
                         if (hasStatusBar) {
-                            return <PlaceholderComponent id={DefaultComponentNames.ScaleDisplay} locale={this.props.config.locale} componentProps={SCALE_DISPLAY_PROPS} />;
+                            return <PlaceholderComponent id={DefaultComponentNames.ScaleDisplay} locale={this.props.config.locale} />;
                         }
                     })()}
                     {(() => {
                         if (hasStatusBar) {
-                            return <PlaceholderComponent id={DefaultComponentNames.SelectedFeatureCount} locale={this.props.config.locale} componentProps={SELECTED_FEATURE_COUNT_PROPS} />;
+                            return <PlaceholderComponent id={DefaultComponentNames.SelectedFeatureCount} locale={this.props.config.locale} />;
                         }
                     })()}
-                    <PlaceholderComponent id={DefaultComponentNames.PoweredByMapGuide} locale={this.props.config.locale} componentProps={PBMG_PROPS} />
+                    <PlaceholderComponent id={DefaultComponentNames.PoweredByMapGuide} locale={this.props.config.locale} />
                 </div>
             })()}
             {(() => {